CODE RUB:  Changed the exception type from CommentValidationException to CommentDependencyValidationException and moved the unit tests to the CommentServiceTests.Exceptions.Add and CommentServiceTests.Exceptions.Modify files
