+++ Taarafo.Core.Tests.Unit/Services/Foundations/Comments/CommentServiceTests.Validation.Modify.cs
@@ -370,52 +370,5 @@ namespace Taarafo.Core.Tests.Unit.Services.Foundations.Comments
             this.dateTimeBrokerMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
         }
-
-        [Fact]
-        public async void ShouldThrowValidationExceptionOnModifyIfReferenceErrorOccursAndLogItAsync()
-        {
-            // given
-            int minuteInPast = GetRandomNegativeNumber();
-            DateTimeOffset randomDate = GetRandomDateTimeOffset();
-            Comment randomComment = CreateRandomModifyComment(randomDate.AddMinutes(minuteInPast));
-            Comment inputComment = randomComment.DeepClone();
-            inputComment.UpdatedDate = randomDate;
-            Comment storageComment = randomComment;
-            string randomMessage = GetRandomMessage();
-            string exceptionMessage = randomMessage;
-            
-            var foreignKeyConstraintConflictException = 
-                new ForeignKeyConstraintConflictException(exceptionMessage);
-
-            var invalidCommentReferenceException =
-                new InvalidCommentReferenceException(foreignKeyConstraintConflictException);
-
-            var expectedCommentValidationException =
-                new CommentValidationException(invalidCommentReferenceException);
-
-            this.dateTimeBrokerMock.Setup(broker =>
-                broker.GetCurrentDateTimeOffset())
-                    .Throws(foreignKeyConstraintConflictException);
-
-            // when
-            ValueTask<Comment> modifyCommentTask =
-                this.commentService.ModifyCommentAsync(inputComment);
-
-            // then
-            await Assert.ThrowsAsync<CommentValidationException>(() =>
-                modifyCommentTask.AsTask());
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTimeOffset(),
-                    Times.Once);
-
-            this.loggingBrokerMock.Verify(broker =>
-                broker.LogError(It.Is(SameExceptionAs(expectedCommentValidationException))),
-                    Times.Once);
-
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.storageBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-        }
     }
 }
