+++ Taarafo.Core.Tests.Unit/Services/Foundations/Comments/CommentServiceTests.Validations.Add.cs
@@ -210,51 +210,6 @@ namespace Taarafo.Core.Tests.Unit.Services.Foundations.Comments
             this.dateTimeBrokerMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
             this.storageBrokerMock.VerifyNoOtherCalls();
-        }
-
-        [Fact]
-        public async void ShouldThrowValidationExceptionOnAddIfReferenceErrorOccursAndLogItAsync()
-        {
-            // given
-            DateTimeOffset randomDateTime =
-                GetRandomDateTimeOffset();
-
-            Comment randomComment = CreateRandomComment(randomDateTime);
-            Comment someComment = randomComment;
-            string randomMessage = GetRandomMessage();
-            string exceptionMessage = randomMessage;
-            var foreignKeyConstraintConflictException = new ForeignKeyConstraintConflictException(exceptionMessage);
-
-            var invalidCommentReferenceException =
-                new InvalidCommentReferenceException(foreignKeyConstraintConflictException);
-
-            var expectedCommentValidationException =
-                new CommentValidationException(invalidCommentReferenceException);
-
-            this.dateTimeBrokerMock.Setup(broker =>
-                broker.GetCurrentDateTimeOffset())
-                .Throws(foreignKeyConstraintConflictException);
-
-            // when
-            ValueTask<Comment> addCommentTask =
-                this.commentService.AddCommentAsync(someComment);
-
-            // then
-            await Assert.ThrowsAsync<CommentValidationException>(() =>
-                addCommentTask.AsTask());
-
-            this.dateTimeBrokerMock.Verify(broker =>
-                broker.GetCurrentDateTimeOffset(),
-                    Times.Once());
-
-            this.loggingBrokerMock.Verify(broker =>
-               broker.LogError(It.Is(SameExceptionAs(expectedCommentValidationException))),
-                    Times.Once);
-
-            this.storageBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-        }
-
+        }       
     }
 }
